+++ lib/flutter_cache.dart
@@ -1,8 +1,0 @@
-library flutter_cache;
-
-export 'src/cache_value.dart';
-export 'src/cache_value_def.dart';
-export 'src/cache_value_notifier.dart';
-export 'src/cache_value_with_default.dart';
-export 'src/cache_value_with_default_def.dart';
-export 'src/cache_value_with_default_notifier.dart';