+++ lib/flutter_prefs.dart
@@ -1,0 +1,8 @@
+library flutter_prefs;
+
+export 'src/prefs_value.dart';
+export 'src/prefs_value_def.dart';
+export 'src/prefs_value_notifier.dart';
+export 'src/prefs_value_with_default.dart';
+export 'src/prefs_value_with_default_def.dart';
+export 'src/prefs_value_with_default_notifier.dart';