+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* Flutter cache initial commit.
+* Flutter prefs initial commit.